requirements.txt
fastapi
uvicorn[standard]
httpx
pyyaml
tree-sitter>=0.23
tree-sitter-python
tree-sitter-javascript
tree-sitter-typescript
tree-sitter-go
networkx
sqlite-vec
tiktoken
gitpython
watchfiles
ollama
cachetools
